configure.in:
  Auto merged
include/my_pthread.h:
  Auto merged
libmysql/libmysql.c:
  Auto merged
server-tools/instance-manager/guardian.cc:
  Auto merged
sql/field.cc:
  Auto merged
sql/field.h:
  Auto merged
sql/item_func.cc:
  Auto merged
sql/item_strfunc.cc:
  Auto merged
sql/item_sum.cc:
  Auto merged
sql/mysql_priv.h:
  Auto merged
sql/mysqld.cc:
  Auto merged
sql/set_var.cc:
  Auto merged
sql/sql_cache.cc:
  Auto merged
sql/sql_lex.h:
  Auto merged
sql/sql_plugin.cc:
  Auto merged
sql/sql_trigger.cc:
  Auto merged
sql/sql_yacc.yy:
  Auto merged
sql/table.cc:
  Auto merged
storage/archive/ha_archive.cc:
  Auto merged
storage/csv/ha_tina.cc:
  Auto merged
storage/federated/ha_federated.cc:
  Auto merged
sql/item_create.cc:
  manual merge

+++ mysys/my_vle.c
@@ -52,7 +52,7 @@ my_vle_encode(byte* out, my_size_t max, ulong n)
 
   do
   {
-    *ptr++= (n & 0x7F);
+    *ptr++= (byte) (n & 0x7F);
     n>>= 7;
   }
   while (n > 0);
